+++ backend/golang/src/go.mod
@@ -2,16 +2,14 @@ module gymlink
 
 go 1.23.5
 
-require github.com/go-sql-driver/mysql v1.9.3
+require (
+	github.com/go-sql-driver/mysql v1.9.3
+	github.com/golang-migrate/migrate v3.5.4+incompatible
+	github.com/golang-migrate/migrate/v4 v4.19.0
+)
 
 require (
 	filippo.io/edwards25519 v1.1.0 // indirect
-	github.com/golang-migrate/migrate v3.5.4+incompatible // indirect
-	github.com/golang-migrate/migrate/v4 v4.19.0 // indirect
-	github.com/google/go-github/v39 v39.2.0 // indirect
-	github.com/google/go-querystring v1.1.0 // indirect
 	github.com/hashicorp/errwrap v1.1.0 // indirect
 	github.com/hashicorp/go-multierror v1.1.1 // indirect
-	golang.org/x/crypto v0.36.0 // indirect
-	golang.org/x/oauth2 v0.27.0 // indirect
 )
