--- backend/golang/src/internal/db/migration/000002_create_initial_tables.down.sql
@@ -1,1 +1,0 @@
-DROP TABLE IF EXISTS "users";
