+++ backend/golang/src/internal/dbase/migration/000002_create_initial_tables.up.sql
@@ -1,0 +1,1 @@
+
