+++ backend/golang/src/internal/dbase/migration/000003_create_characters_table.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS `characters`;
