+++ backend/golang/src/internal/dbase/migration/000003_create_characters_table.up.sql
@@ -1,0 +1,9 @@
+CREATE TABLE characters (
+    `id` INT NOT NULL PRIMARY KEY,
+    `name` VARCHAR(255) NOT NULL,   
+    `image_url` VARCHAR(255) NOT NULL,
+    `level` INT NOT NULL,
+    `limit_point` INT NOT NULL,
+    `created_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    `updated_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
+);
