+++ backend/golang/src/internal/dbase/migration/000004_create_users_table.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS `users`;
