+++ backend/golang/src/internal/dbase/migration/000004_create_users_table.up.sql
@@ -1,9 +1,10 @@
 CREATE TABLE users (
     `id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `character_id` INT NOT NULL,
     `firebase_uid` VARCHAR(255) NOT NULL,
     `name` VARCHAR(255) NOT NULL,
     `email` VARCHAR(255) NOT NULL UNIQUE,
-    `avatar_url` varchar(255),
     `created_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    `updated_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
+    `updated_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+    FOREIGN KEY (character_id) REFERENCES characters(id)
 );
