+++ backend/golang/src/internal/dbase/migration/000005_create_profiles_table.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS `profiles`;
