+++ backend/golang/src/internal/dbase/migration/000006_create_exercise_records_table.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS `exercise_records`;
