+++ backend/golang/src/internal/dbase/migration/000006_create_exercise_records_table.up.sql
@@ -1,0 +1,10 @@
+CREATE TABLE exercise_records (
+    `id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `user_id` INT NOT NULL,
+    `exercise_time` INT NOT NULL,
+    `exercise_date` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    `comment` VARCHAR(1000),
+    `created_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    `updated_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+    FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
+);
