+++ backend/golang/src/internal/dbase/migration/000007_create_user_likes_table.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS `user_likes`;
