+++ backend/golang/src/internal/dbase/migration/000008_create_follows_table.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS `follows`;
