+++ backend/golang/src/internal/dbase/migration/000008_create_follows_table.up.sql
@@ -1,0 +1,10 @@
+CREATE TABLE follows (
+    `follower_id` INT NOT NULL,
+    `followed_id` INT NOT NULL,
+    `created_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    `updated_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+    FOREIGN KEY (follower_id) REFERENCES users(id) ON DELETE CASCADE,
+    FOREIGN KEY (followed_id) REFERENCES users(id) ON DELETE CASCADE,    
+    PRIMARY KEY(follower_id,followed_id),
+    CHECK(follower_id <> followed_id)
+);
